persist local module
